+++ fx/src/cont/ab/eclectic.rs
@@ -3,9 +3,11 @@
 //! date: 2023/02/12 22:33:08 Sunday
 //! brief: Eclectic
 
+use std::hash::Hash;
+
 use arrow2::datatypes::DataType;
 
-use crate::cont::ab::private;
+use crate::cont::private;
 use crate::{FxError, FxResult};
 
 use super::FxSeq;
@@ -21,43 +23,52 @@ pub trait Eclectic: private::InnerEclectic + Clone {
     }
 
     fn is_arr(&self) -> bool {
-        private::InnerEclectic::is_arr(self)
+        Self::Seq::is_arr()
     }
 
     fn is_vec(&self) -> bool {
-        private::InnerEclectic::is_vec(self)
+        Self::Seq::is_vec()
     }
 
     fn width(&self) -> usize {
-        private::InnerEclectic::width(self)
+        self.ref_sequences().iter().count()
     }
 
     fn lens(&self) -> Vec<usize> {
-        private::InnerEclectic::lens(self)
+        self.ref_sequences()
+            .iter()
+            .map(|s| s.len())
+            .collect::<Vec<_>>()
     }
 
+    // if `lens()` is empty, return `None`
     fn max_len(&self) -> Option<usize> {
-        private::InnerEclectic::max_len(self)
+        self.lens().iter().max().cloned()
     }
 
+    // if `lens()` is empty, return `None`
     fn min_len(&self) -> Option<usize> {
-        private::InnerEclectic::min_len(self)
+        self.lens().iter().min().cloned()
     }
 
     fn is_lens_same(&self) -> bool {
-        private::InnerEclectic::is_lens_same(self)
+        let l = self.lens();
+
+        l.first()
+            .map(|first| l.iter().all(|x| x == first))
+            .unwrap_or(true)
     }
 
     fn is_empty(&self) -> bool {
-        private::InnerEclectic::is_empty(self)
+        self.ref_sequences().is_empty()
     }
 
     fn data_types(&self) -> Vec<&DataType> {
-        private::InnerEclectic::data_types(self)
+        self.ref_sequences().iter().map(|e| e.data_type()).collect()
     }
 
     fn data_types_match<T: Eclectic>(&self, d: &T) -> bool {
-        private::InnerEclectic::data_types_match(self, d)
+        self.width() == d.width() && self.data_types() == d.data_types()
     }
 
     fn sequences(&self) -> &[Self::Seq] {
@@ -66,6 +77,10 @@ pub trait Eclectic: private::InnerEclectic + Clone {
 
     fn into_sequences(self) -> Vec<Self::Seq> {
         self.take_sequences()
+    }
+
+    fn check_nulls(&self) -> Vec<bool> {
+        self.ref_sequences().iter().map(|s| s.has_null()).collect()
     }
 
     fn slice() {
@@ -98,3 +113,48 @@ pub trait Eclectic: private::InnerEclectic + Clone {
 }
 
 impl<T> Eclectic for T where T: private::InnerEclectic + Clone {}
+
+// ================================================================================================
+// EclecticCollection
+// ================================================================================================
+
+pub trait EclecticCollection<I, C>: private::InnerEclecticCollection<I, C>
+where
+    I: Hash,
+    C: Eclectic,
+{
+    fn length(&self) -> usize {
+        self.ref_container().len()
+    }
+
+    fn width(&self) -> usize {
+        self.ref_schema().fields.len()
+    }
+
+    fn size(&self) -> (usize, usize) {
+        (self.length(), self.width())
+    }
+
+    fn is_empty(&self) -> bool {
+        self.ref_container().is_empty()
+    }
+
+    fn data_types(&self) -> Vec<&DataType> {
+        self.ref_schema()
+            .fields
+            .iter()
+            .map(|f| f.data_type())
+            .collect::<Vec<_>>()
+    }
+
+    fn data_types_check(&self, c: &C) -> bool {
+        self.width() == c.width() && self.data_types() == c.data_types()
+    }
+
+    fn data_types_match<T>(&self, d: &T) -> bool
+    where
+        T: EclecticCollection<I, C>,
+    {
+        self.width() == d.width() && self.data_types() == d.data_types()
+    }
+}
