+++ fx/src/cont/mod.rs
@@ -12,6 +12,7 @@ pub mod nullopt;
 pub mod parcel;
 pub mod table;
 
+pub use ab::*;
 pub use batch::*;
 pub use bundle::*;
 pub use cvt::*;
