+++ fx/src/macros.rs
@@ -346,7 +346,7 @@ macro_rules! impl_sql_meta {
             //     sql: &'a str,
             // ) -> ::futures::future::BoxFuture<'a, FxResult<FxGrid>>
             // where
-            //     D: Send + $crate::cont::ab::FxChunkingRowBuilderGenerator<FxGrid>,
+            //     D: Send + $crate::cont::FxChunkingRowBuilderGenerator<FxGrid>,
             //     D: From<Self::Row>,
             // {
             //     let q = async move {
